# docs/swagger.yaml
definitions:
...
  go-file-server_internal_service.CancelReq:
    properties:
      cancel:
        description: optional explicit cancel flag
        type: boolean
      opId:
        type: string
    required:
    - opId
    type: object
  go-file-server_internal_service.CopyReq:
    properties:
      destDir:
        type: string
      opId:
        description: 'Optional: client provided operation ID'
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - destDir
    - sources
    type: object
  go-file-server_internal_service.CreateFolderReq:
    properties:
      dir:
        description: Target directory path (defaults to root if empty).
        type: string
      folderName:
        type: string
      opId:
        type: string
    required:
    - folderName
    type: object
  go-file-server_internal_service.DeleteReq:
    properties:
      opId:
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - sources
    type: object
  go-file-server_internal_service.MoveReq:
    properties:
      destDir:
        type: string
      opId:
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - destDir
    - sources
    type: object
  go-file-server_internal_service.PropertiesContains:
    properties:
      files:
        type: integer
      folders:
        type: integer
    type: object
  go-file-server_internal_service.PropertiesReq:
    properties:
      sources:
        items:
          type: string
        type: array
    required:
    - sources
    type: object
  go-file-server_internal_service.PropertiesResponse:
    properties:
      accessedAt:
        type: string
      contains:
        $ref: '#/definitions/go-file-server_internal_service.PropertiesContains'
      createdAt:
        type: string
      location:
        type: string
      modifiedAt:
        type: string
      name:
        description: null if multiple
        type: string
      totalSizeBytes:
        type: integer
      type:
        description: '"file", "directory", or "multiple"'
        type: string
    type: object
  go-file-server_internal_service.RenameReq:
    properties:
      newName:
        type: string
      opId:
        type: string
      source:
        type: string
    required:
    - newName
    - source
    type: object
...
  internal_service.CheckDuplicateReq:
    properties:
      destDir:
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - destDir
    - sources
    type: object
  internal_service.CheckUploadDuplicateReq:
    properties:
      destDir:
        type: string
      files:
        items:
          $ref: '#/definitions/internal_service.UploadFileDetail'
        type: array
    required:
    - destDir
    - files
    type: object
...
  internal_service.UploadFileDetail:
    properties:
      modifiedAt:
        type: string
      path:
        description: relative path including filename
        type: string
      size:
        type: integer
    type: object
...
host: localhost:3333
schemes:
  - http
info:
  contact:
    name: GoNet Drive Support
    url: https://github.com
  description: GoNet Drive is a self-hosted cloud file storage and sharing service.
  termsOfService: https://example.com/terms
  title: GoNet Drive API
  version: "1.0"
paths:
...
  /api/share/file/copy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Copy request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CopyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Copy Files in Share
      tags:
      - Share Files
  /api/share/file/create-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create folder request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CreateFolderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Create Folder in Share
      tags:
      - Share Files
  /api/share/file/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Soft Delete Share Files
      tags:
      - Share Files
  /api/share/file/delete-permanent:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Permanent Delete Share Files
      tags:
      - Share Files
  /api/share/file/download:
    get:
      description: Download one or more files/directories from a shared path. Single
        files are served directly; multiple items are packaged as ZIP.
      parameters:
      - collectionFormat: csv
        description: Source file paths
        in: query
        items:
          type: string
        name: source
        required: true
        type: array
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - application/octet-stream
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Download Share Files
      tags:
      - Share Files
...
  /api/share/file/move:
    post:
      consumes:
      - application/json
      parameters:
      - description: Move request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.MoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Move Files in Share
      tags:
      - Share Files
  /api/share/file/properties:
    post:
      consumes:
      - application/json
      description: Get metadata for files/directories in a shared path.
      parameters:
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      - description: Properties request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.PropertiesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-file-server_internal_service.PropertiesResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Share File Properties
      tags:
      - Share Files
  /api/share/file/rename:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rename request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.RenameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Rename File in Share
      tags:
      - Share Files
  /api/share/file/upload-chunk:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file chunk to a shared path. Requires modify authority.
      parameters:
      - description: Upload identifier
        in: formData
        name: identifier
        required: true
        type: string
      - description: 'Chunk status: start, uploading, end, cancel'
        in: formData
        name: status
        required: true
        type: string
      - description: Target filename
        in: formData
        name: filename
        required: true
        type: string
      - description: Destination directory
        in: formData
        name: destination
        required: true
        type: string
      - description: Chunk number (1-indexed)
        in: formData
        name: chunkNumber
        required: true
        type: integer
      - description: Total number of chunks
        in: formData
        name: totalChunks
        required: true
        type: integer
      - description: Chunk checksum
        in: formData
        name: checksum
        type: string
      - description: The chunk file data
        in: formData
        name: chunk
        required: true
        type: file
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Upload to Share
      tags:
      - Share Files
...
  /api/user/files/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a running or queued file operation by its operation ID.
      parameters:
      - description: Cancel request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CancelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Cancel Operation
      tags:
      - Files
  /api/user/files/check-duplicates:
    post:
      consumes:
      - application/json
      description: Check for name collision between source and destination paths.
      parameters:
      - description: Check duplicates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_service.CheckDuplicateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Check Duplicates
      tags:
      - Files
  /api/user/files/check-upload-duplicates:
    post:
      consumes:
      - application/json
      description: Check for file name collision when uploading files to a destination.
      parameters:
      - description: Check upload duplicates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_service.CheckUploadDuplicateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Check Upload Duplicates
      tags:
      - Files
  /api/user/files/copy:
    post:
      consumes:
      - application/json
      description: Start an asynchronous copy operation for files/directories.
      parameters:
      - description: Copy request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CopyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Copy Files
      tags:
      - Files
  /api/user/files/create-folder:
    post:
      consumes:
      - application/json
      description: Create a new directory.
      parameters:
      - description: Create folder request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CreateFolderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Create Folder
      tags:
      - Files
  /api/user/files/delete:
    post:
      consumes:
      - application/json
      description: Move files/directories to the recycle bin (.cloud_delete).
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Soft Delete Files
      tags:
      - Files
  /api/user/files/delete-permanent:
    post:
      consumes:
      - application/json
      description: Permanently delete files/directories (bypass recycle bin).
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Permanent Delete
      tags:
      - Files
  /api/user/files/download:
    get:
      description: Download one or more files/directories. Single files are served
        directly; multiple items are packaged as a ZIP.
      parameters:
      - collectionFormat: csv
        description: Source file paths
        in: query
        items:
          type: string
        name: source
        required: true
        type: array
      produces:
      - application/octet-stream
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Download Files
      tags:
      - Files
  /api/user/files/file-list:
    get:
      description: List files and directories at the specified path. Supports sorting
        by name, size, or modified time.
      parameters:
      - description: 'Directory path (default: /)'
        in: query
        name: path
        type: string
      - description: 'Sort by: name, size, modified (default: name)'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc, desc (default: asc)'
        in: query
        name: order
        type: string
      - description: 'Show hidden files: true, false (default: false)'
        in: query
        name: showHidden
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List Files
      tags:
      - Files
  /api/user/files/file-test:
    get:
      description: Simple health check for the files API module.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: File API Health Check
      tags:
      - Files
  /api/user/files/move:
    post:
      consumes:
      - application/json
      description: Start an asynchronous move operation for files/directories.
      parameters:
      - description: Move request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.MoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Move Files
      tags:
      - Files
  /api/user/files/properties:
    post:
      consumes:
      - application/json
      description: Get metadata properties for files or directories.
      parameters:
      - description: Properties request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.PropertiesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-file-server_internal_service.PropertiesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: File Properties
      tags:
      - Files
  /api/user/files/rename:
    post:
      consumes:
      - application/json
      description: Rename a file or directory.
      parameters:
      - description: Rename request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.RenameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Rename File
      tags:
      - Files
  /api/user/files/storage-usage:
    get:
      description: Get current storage usage and limit.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Storage Usage
      tags:
      - Files
  /api/user/files/upload-chunk:
    post:
      consumes:
      - multipart/form-data
      description: Upload a chunk of a file. Supports resumable uploads with start/uploading/end/cancel
        status.
      parameters:
      - description: Upload identifier
        in: formData
        name: identifier
        required: true
        type: string
      - description: 'Chunk status: start, uploading, end, cancel'
        in: formData
        name: status
        required: true
        type: string
      - description: Target filename
        in: formData
        name: filename
        required: true
        type: string
      - description: Destination directory
        in: formData
        name: destination
        required: true
        type: string
      - description: Chunk number (1-indexed)
        in: formData
        name: chunkNumber
        required: true
        type: integer
      - description: Total number of chunks
        in: formData
        name: totalChunks
        required: true
        type: integer
      - description: Chunk checksum (CRC32/MD5/SHA256)
        in: formData
        name: checksum
        type: string
      - description: The chunk file data
        in: formData
        name: chunk
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties: true
            type: object
        "507":
          description: Insufficient Storage
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Upload Chunk
      tags:
      - Files
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token stored as a cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"